requirements.txt
fastapi
uvicorn[standard]
pika
elasticsearch
pyspark
python-docx
pymupdf

